backend/requirements.txt
websockets>=12.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
PyGithub>=2.1.1
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pytesseract>=0.3.0
scipy
scikit-learn
ultralytics>=8.0.0
